rename chatchatbotwriting to chatbotwriting
